Add files via upload
CONCLUSION

Implementing the project is very much customer friendly, easy operation at finger tips using
AADHAR CARD which is a must. Use of AADHAR card technology makes the system
secure as it gives unique identification to the customer. Once after installing, maintenance is
very easy. Long queues, time and efforts is saved. The main aim of the project is to give the
business, substantial savings in labor, time and staff training costs. The software coding in
Embedded C code gives the required flexibility to design the transactions as per customer's
requirement and convenience. Due to automation, fraudulent transactions are eliminated.
Following are the advantages of the AADHAR based self service petrol pump
Improves the use of GSM module in digital transactions

Increases the efficiency of the services of petrol pump (especially in late night
hours)
Enhances the best utilization of TIME
People below 18 years of age will not able to access the petrol pump service
Increases the overall profit of the petrol pump